+++ clikt/src/mingwX64Main/kotlin/com/github/ajalt/clikt/output/defaultCliktConsole.kt
@@ -1,0 +1,3 @@
+package com.github.ajalt.clikt.output
+
+actual fun defaultCliktConsole(): CliktConsole = defaultNativeCliktConsole()